+++ apps/cowswap-frontend/src/legacy/components/Copy/CopyMod.tsx
@@ -2,11 +2,11 @@ import React, { MouseEvent } from 'react'
 
 import { useCopyClipboard } from '@cowprotocol/common-hooks'
 import { UI } from '@cowprotocol/ui'
+import { LinkStyledButton } from '@cowprotocol/ui'
 
 import { Trans } from '@lingui/react/macro'
 import { CheckCircle, Copy } from 'react-feather'
 import styled, { DefaultTheme, StyledComponentProps } from 'styled-components/macro'
-import { LinkStyledButton } from 'theme'
 
 import { TransactionStatusText } from 'legacy/components/Copy/index'
 
@@ -54,12 +54,13 @@ interface CopyHelperProps
   children?: React.ReactNode
   clickableLink?: boolean
   copyIconWidth?: string
+  hideCopiedLabel?: boolean
 }
 
 // TODO: Add proper return type annotation
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export default function CopyHelper(props: CopyHelperProps) {
-  const { toCopy, children, clickableLink, copyIconWidth, ...rest } = props
+  const { toCopy, children, clickableLink, copyIconWidth, hideCopiedLabel = false, ...rest } = props
   const [isCopied, setCopied] = useCopyClipboard()
 
   // TODO: Add proper return type annotation
@@ -78,11 +79,13 @@ export default function CopyHelper(props: CopyHelperProps) {
             isCopied={isCopied} // mod
           >
             <CheckCircleIconWrapper size={'16'} />
-            <TransactionStatusText
-              isCopied={isCopied} // mod
-            >
-              <Trans>Copied</Trans>
-            </TransactionStatusText>
+            {!hideCopiedLabel ? (
+              <TransactionStatusText
+                isCopied={isCopied} // mod
+              >
+                <Trans>Copied</Trans>
+              </TransactionStatusText>
+            ) : null}
           </TransactionStatusText>
         ) : (
           <TransactionStatusText>